Move EVSE example parser functions to separate file
