+++ Examples/cli/individual/EV_SB/getEvseCurrentlyAppliedControlSettings.js
@@ -5,7 +5,9 @@ const {
     EMCB_UDP_EVENT_DEVICE_DISCOVERED
 } = require('../../../..'); // If running this example somewhere outside of a `git clone` of the `emcb-udp-master` module, replace with `require("emcb-udp-master")`
 
-const UDPKeys = require("../../../_config.js")
+const UDPKeys = require("../../../_config.js");
+
+var parsers = require('./evseExampleParsers.js');
 
 var EMCBs = new EmcbUDPbroadcastMaster({
     broadcastUDPKey : UDPKeys.broadcast,
@@ -18,28 +20,6 @@ var deviceIP = argv._[1];
 console.log("Establishing connection to device " + deviceID + " at ip: " + deviceIP);
 var device = EMCBs.createDevice(deviceID, deviceIP, true);
 
-function parseGetAppliedEvseControlSettings(response){
-    // clone the object
-    var response = Object.assign({}, response); // shallow copy
-    delete response.device;
-    delete response.raw;
-
-    response.enabled = (response.enabled == 1 ? true : (response.enabled == 0 ? false : null));
-    response.authorized = (response.authorized == 1 ? true : (response.authorized == 0 ? false : null));
-
-    // 0 means no software limit set, but hardware is still limited to 32 amps
-    if (response.maxCurrentAmps == 0){
-        response.maxCurrentAmps = 32;
-    }
-
-    if (response.maxEnergyWatts == 0){
-        response.maxEnergyWatts = "none";
-    }
-
-    return response;
-}
-
-
 EMCBs.on(EMCB_UDP_EVENT_DEVICE_DISCOVERED, (data) => {
     console.log('Device connected from address: ' + data.device.ipAddress);
     console.log("Sending: Get EVSE currently applied control settings");
@@ -48,7 +28,7 @@ EMCBs.on(EMCB_UDP_EVENT_DEVICE_DISCOVERED, (data) => {
         .then(data => {
             for (const [ip, response] of Object.entries(data.responses)) {
                 console.log("Response from " + ip + ":");
-                console.log(parseGetAppliedEvseControlSettings(response));
+                console.log(parsers.parseGetAppliedEvseControlSettings(response));
             }
         })
         .catch(err => {
